Refactored Groups Router to include subrouter from each sub path endpoint.
