+++ api/groups/groupActivities/groupActivitiesModel.js
@@ -1,0 +1,9 @@
+const db = require('../../../data/dbConfig.js');
+
+module.exports = {
+
+    getGroupActivity: function(groupId) {
+        return db('activities').where({ groupId })
+    },
+    
+};
