+++ api/groups/groupOwners/groupOwnersModel.js
@@ -1,0 +1,18 @@
+const db = require('../../../data/dbConfig.js');
+
+module.exports = {
+
+    getGroupOwners: function(groupId) {
+
+        return db('usersGroupsOwnership')
+            .select(
+                'usersGroupsOwnership.groupId', 
+                'usersGroupsOwnership.userId', 
+                'users.displayName',
+            )
+            .where({ groupId })
+            .join('users', 'usersGroupsOwnership.userId', 'users.id')
+
+    },
+    
+};
