+++ api/groups/groupsModel.spec.js
@@ -34,54 +34,4 @@ describe('groupModel', () => {
 
     });
 
-    describe('getGroupMembers()', () => {
-
-        it('should return list all members of group by specified group ID', async () => {
-            const res = await groupModel.getGroupMembers(id);
-            expect(res).toHaveLength;
-            expect(res[0].groupId).toBe(id);
-            expect(res[0].userId).toBeDefined();
-            expect(res[0].displayName).toBeDefined();
-
-        });
-
-    });
-
-    describe('getGroupOwners()', () => {
-
-        it('should return list all owners of group by specified group ID', async () => {
-            const res = await groupModel.getGroupOwners(id);
-            expect(res).toHaveLength;
-            expect(res[0].groupId).toBe(id);
-            expect(res[0].userId).toBeDefined();
-            expect(res[0].displayName).toBeDefined();
-
-        });
-
-    });
-
-    describe('getGroupActivity()', () => {
-
-        it('should return list all activities by specified group ID', async () => {
-            const res = await groupModel.getGroupActivity(id);
-            expect(res).toHaveLength;
-            expect(res[0].groupId).toBe(id);
-            expect(res[0].userId).toBeDefined();
-            expect(res[0].activity).toBeDefined();
-
-        });
-
-    });
-
-    describe('getGroupCallStatus()', () => {
-
-        it('should return call status by specified group ID', async () => {
-            const res = await groupModel.getGroupCallStatus(id);
-            expect(res.callStatus).toBeDefined();
-            expect(res.callStatus).toBe(1 || 0);
-
-        });
-
-    });
-
 });
